Clean up demographics
